PHASE 2: Fix BigDecimal TypeConverters — restore build
- Added BigDecimalConverters.kt for Room database
- Updated AppDatabase.kt with TypeConverters annotation
- Converted AccountEntity, BudgetEntity, TransactionEntity, GoalEntity from Double to BigDecimal
- Converted Account, Budget, Transaction, Goal domain models from Double to BigDecimal
- Updated mapper files with BigDecimal imports
- Bumped database version to 14

Closes P0-001, P0-002, P0-003,

+++ app/src/main/java/com/taskmanager/domain/model/Transaction.kt
@@ -1,5 +1,6 @@
 package com.taskmanager.domain.model
 
+import java.math.BigDecimal
 import java.time.Instant
 
 enum class TransactionType {
@@ -8,7 +9,7 @@ enum class TransactionType {
 
 data class Transaction(
     val id: Long? = null,
-    val amount: Double,
+    val amount: BigDecimal,
     val type: TransactionType,
     val currency: String = "RUB",
     val categoryId: Long? = null,
@@ -17,7 +18,7 @@ data class Transaction(
     val note: String? = null,
     // Transfer-specific fields
     val toAccountId: Long? = null,
-    val destinationAmount: Double? = null,
+    val destinationAmount: BigDecimal? = null,
     val destinationCurrency: String? = null,
     val createdAt: Instant = Instant.now(),
     val updatedAt: Instant = Instant.now()
